mba-ai/lang-chain/classes/1-fundamentos/2-initi-chat-model.py
from langchain.chat_models import init_chat_model
from dotenv import load_dotenv
load_dotenv()

gemini = init_chat_model(model="gemini-2.5-flash", model_provider="google_genai")
answer = gemini.invoke("Hello World")
print(answer.content)